Relax pins for plugins
Partially addresses https://github.com/conda-forge/dbt-feedstock/pull/60#issuecomment-957727034 by relaxing the pinning of the plugins to the major/minor versions.

This recipe may still be problematic if for example `dbt-core` bumps the patch version while there are no corresponding patch versions for the other plugins. I'm not sure how to solve this.